requirements.txt
fastapi
uvicorn
pdfplumber
sentence-transformers
faiss-cpu
ollama
python-multipart
numpy
